+++ loqusdb/build_models/variant.py
@@ -15,49 +15,59 @@ Position = namedtuple("Position", "chrom pos")
 # These are coordinate for the pseudo autosomal regions in GRCh37
 
 
-def check_par(chrom, pos, genome_build=None):
+def check_par(chrom, pos, genome_build):
     """Check if a coordinate is in the PAR region
 
     Args:
         chrom(str)
         pos(int)
+        genome_build(str): Genome build. Ex. GRCh37 or GRCh38
 
     Returns:
         par(bool)
     """
-    if genome_build is None:
-        genome_build = GRCH37
     return any(
         pos >= interval[0] and pos <= interval[1] for interval in PAR[genome_build].get(chrom, [])
     )
 
 
-def get_variant_id(variant):
-    """Get a variant id on the format chrom_pos_ref_alt"""
+def get_variant_id(variant, keep_chr_prefix=None):
+    """Get a variant id on the format chrom_pos_ref_alt
+
+    Args:
+        variant (cyvcf2.Variant)
+        keep_chr_prefix(bool): Retain chr/CHR/Chr prefix when present
+
+    Returns:
+        variant (models.ProfileVariant)
+    """
+
     chrom = variant.CHROM
-    if chrom.lower().startswith("chr"):
-        chrom = chrom[3:]
+    if not keep_chr_prefix:
+        if chrom.lower().startswith("chr"):
+            chrom = chrom[3:]
     return "_".join([str(chrom), str(variant.POS), str(variant.REF), str(variant.ALT[0])])
 
 
-def is_greater(a, b):
+def is_greater(a, b, genome_build):
     """Check if position a is greater than position b
     This will look at chromosome and position.
 
     For example a position where chrom = 2 and pos = 300 is greater than a position where
     chrom = 1 and pos = 1000
 
-    If any of the chromosomes is outside [1-22,X,Y,MT] we can not say which is biggest.
+    If any of the chromosomes is outside [1-22,X,Y,MT] or [chr1-chr22,chrX,chrY,chrM] we can not say which is biggest.
 
     Args:
         a,b(Position)
+        genome_build(str): Genome build. Ex. GRCh37 or GRCh38
 
     Returns:
         bool: True if a is greater than b
     """
 
-    a_chrom = CHROM_TO_INT.get(a.chrom, 0)
-    b_chrom = CHROM_TO_INT.get(b.chrom, 0)
+    a_chrom = CHROM_TO_INT[genome_build].get(a.chrom, 0)
+    b_chrom = CHROM_TO_INT[genome_build].get(b.chrom, 0)
 
     if a_chrom == 0 or b_chrom == 0:
         return False
@@ -68,11 +78,13 @@ def is_greater(a, b):
     return a_chrom == b_chrom and a.pos > b.pos
 
 
-def get_coords(variant):
+def get_coords(variant, keep_chr_prefix, genome_build):
     """Returns a dictionary with position information
 
     Args:
         variant(cyvcf2.Variant)
+        keep_chr_prefix(bool): Retain chr/CHR/Chr prefix when present
+        genome_build(str): Genome build. Ex. GRCh37 or GRCh38
 
     Returns:
         coordinates(dict)
@@ -86,8 +98,9 @@ def get_coords(variant):
         "end": None,
     }
     chrom = variant.CHROM
-    if chrom.startswith(("chr", "CHR", "Chr")):
-        chrom = chrom[3:]
+    if not keep_chr_prefix:
+        if chrom.lower().startswith("chr"):
+            chrom = chrom[3:]
     coordinates["chrom"] = chrom
     end_chrom = chrom
 
@@ -107,8 +120,9 @@ def get_coords(variant):
     if sv_type == "BND":
         other_coordinates = alt.strip("ATCGN").strip("[]").split(":")
         end_chrom = other_coordinates[0]
-        if end_chrom.startswith(("chr", "CHR", "Chr")):
-            end_chrom = end_chrom[3:]
+        if not keep_chr_prefix:
+            if chrom.lower().startswith("chr"):
+                end_chrom = end_chrom[3:]
 
         end = int(other_coordinates[1])
 
@@ -126,7 +140,7 @@ def get_coords(variant):
     end_position = Position(end_chrom, end)
 
     # If 'start' is greater than 'end', switch positions
-    if is_greater(position, end_position):
+    if is_greater(position, end_position, genome_build=genome_build):
         end_chrom = position.chrom
         end = position.pos
 
@@ -148,6 +162,8 @@ def build_variant(
     case_id: Optional[str] = None,
     gq_threshold: Optional[int] = None,
     gq_qual: Optional[bool] = False,
+    keep_chr_prefix: Optional[bool] = False,
+    ignore_gq_if_unset: Optional[bool] = False,
     genome_build: Optional[str] = None,
 ) -> Variant:
     """Return a Variant object
@@ -164,6 +180,9 @@ def build_variant(
         case_id(str): The case id
         gq_threshold(int): Genotype Quality threshold
         gq_qual(bool): Use variant.QUAL for quality instead of GQ
+        keep_chr_prefix(bool): Retain chr/CHR/Chr prefix when present
+        ignore_gq_if_unset(bool): Ignore GQ threshold check for variants that do not have GQ or QUAL set.
+        genome_build(str): Genome build. Ex. GRCh37 or GRCh38
 
     Return:
         formated_variant(models.Variant): A variant dictionary
@@ -176,14 +195,14 @@ def build_variant(
         sv = True
 
     # chrom_pos_ref_alt
-    variant_id = get_variant_id(variant)
+    variant_id = get_variant_id(variant, keep_chr_prefix)
 
     ref = variant.REF
     # ALT is an array in cyvcf2
     # We allways assume splitted and normalized VCFs
     alt = variant.ALT[0]
 
-    coordinates = get_coords(variant)
+    coordinates = get_coords(variant, keep_chr_prefix, genome_build=genome_build)
     chrom = coordinates["chrom"]
     pos = coordinates["pos"]
 
@@ -201,14 +220,15 @@ def build_variant(
             ind_pos = ind_obj["ind_index"]
 
             if gq_qual:
-                gq = 0
+                gq = -1
                 if variant.QUAL:
                     gq = int(variant.QUAL)
 
             if not gq_qual:
                 gq = int(variant.gt_quals[ind_pos])
 
-            if gq_threshold and gq < gq_threshold:
+            # When gq is missing in FORMAT cyvcf2 assigns a score of -1
+            if (gq_threshold and 0 <= gq < gq_threshold) or (gq == -1 and not ignore_gq_if_unset):
                 continue
 
             genotype = GENOTYPE_MAP[variant.gt_types[ind_pos]]
@@ -220,7 +240,7 @@ def build_variant(
                 # If variant in X or Y and individual is male,
                 # we need to check hemizygosity
                 if (
-                    chrom in ["X", "Y"]
+                    chrom in ["X", "Y", "chrX", "chrY"]
                     and ind_obj["sex"] == 1
                     and not check_par(chrom, pos, genome_build=genome_build)
                 ):