+++ loqusdb/commands/load.py
@@ -88,6 +88,20 @@ def validate_profile_threshold(ctx, param, value):
     callback=validate_profile_threshold,
     help="profile hamming distance to store similar individuals (0-1)",
 )
+@click.option(
+    "--snv-gq-only",
+    is_flag=True,
+    default=False,
+    show_default=True,
+    help="Apply GQ threshold only to SNV variants",
+)
+@click.option(
+    "--ignore-gq-if-unset",
+    is_flag=True,
+    default=False,
+    show_default=True,
+    help="Ignore GQ threshold if GQ (or the QUAL field for --qual-gq) is unset in VCF",
+)
 @click.pass_context
 def load(
     ctx,
@@ -104,6 +118,8 @@ def load(
     hard_threshold,
     soft_threshold,
     qual_gq,
+    snv_gq_only,
+    ignore_gq_if_unset,
 ):
     """Load the variants of a case
 
@@ -132,7 +148,7 @@ def load(
 
     adapter = ctx.obj["adapter"]
     genome_build = ctx.obj["genome_build"]
-
+    keep_chr_prefix = ctx.obj["keep_chr_prefix"]
     start_inserting = datetime.now()
 
     try:
@@ -145,12 +161,15 @@ def load(
             skip_case_id=skip_case_id,
             case_id=case_id,
             gq_threshold=gq_threshold,
+            snv_gq_only=snv_gq_only,
+            keep_chr_prefix=keep_chr_prefix,
             qual_gq=qual_gq,
             max_window=max_window,
             profile_file=variant_profile_path,
             hard_threshold=hard_threshold,
             soft_threshold=soft_threshold,
             genome_build=genome_build,
+            ignore_gq_if_unset=ignore_gq_if_unset,
         )
     except (SyntaxError, CaseError, IOError) as error:
         LOG.warning(error)